Add acceleration to player movement.
I feel that this makes the movements feel nicer. From one point of view,
it enables more precise placement with small key taps. This can be used,
for example, to move the pad slightly once, or to take a large jump
followed by a quick corrective movement at the end.

The current dy and ddy parameters seem decent in my limited playtesting.
They're chosen to balance an ability to quickly cross the entire screen
with the ability to fine-tune small movements.