Change name to bounce-beatz.
I've decided to change what this game is fundamentally. It's not pong -
it's something more. Even the basic two-player version has changed
significantly from the original inspiration.

Not that I think this game is amazing or anything. It's still a
day-to-day fun side project. But I'm a little more optimistic that it
may grow up to be something a bit more interesting than I originally
expected.