Partial work toward beatz integration.
I realized that I won't be able to embed .so files at all. This kind of
makes sense, since .love files are meant to be cross-platform. But I
don't really care about being cross-platform right now. I guess I'll
have to work with it. The tricky part is making the beatz library
runnable in all the follwing use cases:
* stand-alone,
* as a regular Lua library, and
* as a Love-based library.