Checkpoint; working on heart-drawing code.
I decided I want to minimze the presence of image files in this repo if
I can help it. So instead of using a heart sprite - I mean something
like a png file - I decided to use a simple mathematical structure to
draw it. It's the overlap of a triangle and the tops of two circles.